requirements.txt
streamlit==1.29.0
langchain-upstage==0.1.8
chromadb==0.4.18
pandas==2.1.4
python-dotenv==1.0.0
tqdm==4.66.1
requests==2.31.0
ollama==0.1.7 